refactor: split shell value type owners
