+++ clients/apple/alan-macos/Models/Shell/TerminalActivityModels.swift
@@ -1,0 +1,418 @@
+import Foundation
+
+enum ShellTabActiveTaskState: String, Codable, Equatable, CaseIterable {
+    case inactive
+    case foregroundCommand = "foreground_command"
+    case alanRunning = "alan_running"
+    case alanPendingYield = "alan_pending_yield"
+    case alanProcess = "alan_process"
+    case unknown
+
+    init(from decoder: Decoder) throws {
+        let container = try decoder.singleValueContainer()
+        let rawValue = try container.decode(String.self)
+        self = Self(rawValue: rawValue) ?? .unknown
+    }
+
+    func encode(to encoder: Encoder) throws {
+        var container = encoder.singleValueContainer()
+        try container.encode(rawValue)
+    }
+
+    var protectsFromPruning: Bool {
+        switch self {
+        case .inactive:
+            return false
+        case .foregroundCommand, .alanRunning, .alanPendingYield, .alanProcess, .unknown:
+            return true
+        }
+    }
+}
+
+enum TerminalActivitySourceKind: String, Codable, Equatable, CaseIterable {
+    case codex
+    case claude
+    case openCode = "open_code"
+    case alan
+    case shell
+    case progress
+    case command
+    case process
+    case unknown
+}
+
+enum TerminalActivityStatus: String, Codable, Equatable, CaseIterable {
+    case needsInput = "needs_input"
+    case failed
+    case paused
+    case progress
+    case running
+    case bell
+    case exited
+    case idle
+    case done
+    case stale
+}
+
+enum TerminalActivityPriority: String, Codable, Equatable, CaseIterable {
+    case passive
+    case active
+    case notable
+    case awaitingUser = "awaiting_user"
+
+    var sidebarPriorityRank: Int {
+        switch self {
+        case .awaitingUser:
+            return 40
+        case .notable:
+            return 30
+        case .active:
+            return 20
+        case .passive:
+            return 10
+        }
+    }
+}
+
+enum TerminalActivityProgressKind: String, Codable, Equatable, CaseIterable {
+    case percent
+    case indeterminate
+    case paused
+    case failed
+}
+
+struct TerminalActivitySource: Codable, Equatable {
+    let kind: TerminalActivitySourceKind
+    let label: String?
+}
+
+struct TerminalActivityProgress: Codable, Equatable {
+    let kind: TerminalActivityProgressKind
+    let percent: Int?
+
+    init(kind: TerminalActivityProgressKind, percent: Int? = nil) {
+        self.kind = kind
+        self.percent = percent.map { min(max($0, 0), 100) }
+    }
+
+    static func percent(_ value: Int) -> TerminalActivityProgress {
+        TerminalActivityProgress(kind: .percent, percent: value)
+    }
+
+    static let indeterminate = TerminalActivityProgress(kind: .indeterminate)
+    static let paused = TerminalActivityProgress(kind: .paused)
+    static let failed = TerminalActivityProgress(kind: .failed)
+}
+
+struct TerminalActivityCommandOutcome: Codable, Equatable {
+    let exitCode: Int?
+    let durationMilliseconds: Int?
+    let commandText: String?
+
+    private enum CodingKeys: String, CodingKey {
+        case exitCode = "exit_code"
+        case durationMilliseconds = "duration_milliseconds"
+        case commandText = "command_text"
+    }
+}
+
+struct TerminalActivityAgentMetadata: Codable, Equatable {
+    let kind: TerminalActivitySourceKind
+    let safeSessionLabel: String?
+    let projectLabel: String?
+    let workingDirectory: String?
+
+    private enum CodingKeys: String, CodingKey {
+        case kind
+        case safeSessionLabel = "safe_session_label"
+        case projectLabel = "project_label"
+        case workingDirectory = "working_directory"
+    }
+}
+
+struct TerminalAgentActivityEvent: Equatable {
+    let agentKind: String
+    let status: String
+    let sessionLabel: String?
+    let projectLabel: String?
+    let workingDirectory: String?
+    let detail: String?
+    let updatedAt: String?
+}
+
+struct TerminalActivityDisplay: Codable, Equatable {
+    let sourceLabel: String
+    let stateLabel: String
+    let detailLabel: String?
+    let paneHint: String?
+
+    var sourceFirstLabel: String {
+        [paneHint, "\(sourceLabel) · \(stateLabel)"]
+            .compactMap { label -> String? in
+                guard let label, !label.isEmpty else { return nil }
+                return label
+            }
+            .joined(separator: " · ")
+    }
+
+    private enum CodingKeys: String, CodingKey {
+        case sourceLabel = "source_label"
+        case stateLabel = "state_label"
+        case detailLabel = "detail_label"
+        case paneHint = "pane_hint"
+    }
+}
+
+struct TerminalActivityFreshness: Codable, Equatable {
+    let updatedAt: String
+    let staleAt: String?
+    let expiresAt: String?
+
+    private enum CodingKeys: String, CodingKey {
+        case updatedAt = "updated_at"
+        case staleAt = "stale_at"
+        case expiresAt = "expires_at"
+    }
+}
+
+struct TerminalActivitySnapshot: Codable, Equatable {
+    private static let iso8601Formatter = ISO8601DateFormatter()
+
+    let source: TerminalActivitySource
+    let status: TerminalActivityStatus
+    let priority: TerminalActivityPriority
+    let progress: TerminalActivityProgress?
+    let command: TerminalActivityCommandOutcome?
+    let agent: TerminalActivityAgentMetadata?
+    let display: TerminalActivityDisplay
+    let freshness: TerminalActivityFreshness
+
+    var isCommandFailure: Bool {
+        source.kind == .command && status == .failed
+    }
+
+    var isSidebarWorthy: Bool {
+        isSidebarWorthy(at: nil, owningTabFocused: false)
+    }
+
+    func isSidebarWorthy(at now: Date?, owningTabFocused: Bool = false) -> Bool {
+        if let now, !isFresh(at: now) {
+            return false
+        }
+        if owningTabFocused,
+           isCommandFailure
+        {
+            return false
+        }
+
+        switch status {
+        case .needsInput, .failed, .paused, .progress, .running, .bell, .exited:
+            return true
+        case .idle, .done, .stale:
+            return false
+        }
+    }
+
+    var sidebarPriorityRank: Int {
+        switch status {
+        case .needsInput:
+            return 70
+        case .failed:
+            return 60
+        case .paused:
+            return 50
+        case .progress:
+            return 40
+        case .running:
+            return 30
+        case .bell, .exited:
+            return 20
+        case .idle, .done, .stale:
+            return 0
+        }
+    }
+
+    func isFresh(at now: Date) -> Bool {
+        if let expiresAt = freshness.expiresAt.flatMap(Self.iso8601Formatter.date(from:)),
+           now >= expiresAt
+        {
+            return false
+        }
+
+        if let staleAt = freshness.staleAt.flatMap(Self.iso8601Formatter.date(from:)),
+           now >= staleAt
+        {
+            return false
+        }
+
+        return true
+    }
+
+    func withPaneHint(_ paneHint: String?) -> TerminalActivitySnapshot {
+        TerminalActivitySnapshot(
+            source: source,
+            status: status,
+            priority: priority,
+            progress: progress,
+            command: command,
+            agent: agent,
+            display: TerminalActivityDisplay(
+                sourceLabel: display.sourceLabel,
+                stateLabel: display.stateLabel,
+                detailLabel: display.detailLabel,
+                paneHint: paneHint
+            ),
+            freshness: freshness
+        )
+    }
+
+    static func primarySidebarActivity(
+        _ activities: [TerminalActivitySnapshot]
+    ) -> TerminalActivitySnapshot? {
+        primarySidebarActivity(activities, now: Date())
+    }
+
+    static func primarySidebarActivity(
+        _ activities: [TerminalActivitySnapshot],
+        now: Date?
+    ) -> TerminalActivitySnapshot? {
+        activities
+            .filter { activity in
+                activity.isSidebarWorthy(at: now)
+            }
+            .max { lhs, rhs in
+                if lhs.sidebarPriorityRank == rhs.sidebarPriorityRank {
+                    if lhs.priority.sidebarPriorityRank == rhs.priority.sidebarPriorityRank {
+                        if lhs.freshness.updatedAt == rhs.freshness.updatedAt {
+                            return lhs.source.kind.rawValue < rhs.source.kind.rawValue
+                        }
+                        return lhs.freshness.updatedAt < rhs.freshness.updatedAt
+                    }
+                    return lhs.priority.sidebarPriorityRank < rhs.priority.sidebarPriorityRank
+                }
+                return lhs.sidebarPriorityRank < rhs.sidebarPriorityRank
+            }
+    }
+
+    static func progressActivity(percent: Int, now: Date) -> TerminalActivitySnapshot {
+        let boundedPercent = min(max(percent, 0), 100)
+        return progressActivity(
+            progress: .percent(boundedPercent),
+            status: .progress,
+            priority: .active,
+            stateLabel: "\(boundedPercent)%",
+            now: now
+        )
+    }
+
+    static func progressActivity(
+        progress: TerminalActivityProgress,
+        status: TerminalActivityStatus,
+        priority: TerminalActivityPriority,
+        stateLabel: String,
+        now: Date
+    ) -> TerminalActivitySnapshot {
+        return TerminalActivitySnapshot(
+            source: TerminalActivitySource(kind: .progress, label: "Progress"),
+            status: status,
+            priority: priority,
+            progress: progress,
+            command: nil,
+            agent: nil,
+            display: TerminalActivityDisplay(
+                sourceLabel: "Progress",
+                stateLabel: stateLabel,
+                detailLabel: nil,
+                paneHint: nil
+            ),
+            freshness: TerminalActivityFreshness(
+                updatedAt: Self.iso8601Formatter.string(from: now),
+                staleAt: Self.iso8601Formatter.string(from: now.addingTimeInterval(15)),
+                expiresAt: nil
+            )
+        )
+    }
+
+    static func commandCompletion(
+        exitCode: Int,
+        now: Date,
+        durationMilliseconds: Int? = nil
+    ) -> TerminalActivitySnapshot {
+        let succeeded = exitCode == 0
+        let status: TerminalActivityStatus = succeeded ? .done : .failed
+        let priority: TerminalActivityPriority = succeeded ? .passive : .notable
+        let stateLabel = succeeded ? "Command succeeded" : "Command failed \(exitCode)"
+        return TerminalActivitySnapshot(
+            source: TerminalActivitySource(kind: .command, label: "Shell"),
+            status: status,
+            priority: priority,
+            progress: nil,
+            command: TerminalActivityCommandOutcome(
+                exitCode: exitCode,
+                durationMilliseconds: durationMilliseconds,
+                commandText: nil
+            ),
+            agent: nil,
+            display: TerminalActivityDisplay(
+                sourceLabel: "Shell",
+                stateLabel: stateLabel,
+                detailLabel: nil,
+                paneHint: nil
+            ),
+            freshness: TerminalActivityFreshness(
+                updatedAt: Self.iso8601Formatter.string(from: now),
+                staleAt: Self.iso8601Formatter.string(from: now.addingTimeInterval(succeeded ? 8 : 30)),
+                expiresAt: nil
+            )
+        )
+    }
+
+    static func bellActivity(now: Date) -> TerminalActivitySnapshot {
+        TerminalActivitySnapshot(
+            source: TerminalActivitySource(kind: .shell, label: "Shell"),
+            status: .bell,
+            priority: .active,
+            progress: nil,
+            command: nil,
+            agent: nil,
+            display: TerminalActivityDisplay(
+                sourceLabel: "Shell",
+                stateLabel: "Bell",
+                detailLabel: nil,
+                paneHint: nil
+            ),
+            freshness: TerminalActivityFreshness(
+                updatedAt: Self.iso8601Formatter.string(from: now),
+                staleAt: nil,
+                expiresAt: Self.iso8601Formatter.string(from: now.addingTimeInterval(8))
+            )
+        )
+    }
+
+    static func processExitedActivity(exitCode: Int?, now: Date) -> TerminalActivitySnapshot {
+        let stateLabel = exitCode.map { "Exited \($0)" } ?? "Exited"
+        return TerminalActivitySnapshot(
+            source: TerminalActivitySource(kind: .process, label: "Process"),
+            status: .exited,
+            priority: .notable,
+            progress: nil,
+            command: TerminalActivityCommandOutcome(
+                exitCode: exitCode,
+                durationMilliseconds: nil,
+                commandText: nil
+            ),
+            agent: nil,
+            display: TerminalActivityDisplay(
+                sourceLabel: "Process",
+                stateLabel: stateLabel,
+                detailLabel: nil,
+                paneHint: nil
+            ),
+            freshness: TerminalActivityFreshness(
+                updatedAt: Self.iso8601Formatter.string(from: now),
+                staleAt: nil,
+                expiresAt: nil
+            )
+        )
+    }
+}
