+++ clients/apple/alan-macos/Services/Shell/ManagedTerminalAccountValidation.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+enum ManagedTerminalAccountIdentifierValidator {
+    static func validate(_ request: ManagedTerminalAccountRequest) -> [ManagedTerminalAccountValidationError] {
+        do {
+            return try ShellCoreFFIAdapter.shared.validateManagedTerminalAccountRequest(request)
+        } catch {
+            return [.coreUnavailable(String(describing: error))]
+        }
+    }
+}
